Implement Authenticatin an Authorization
To implement Authenticatin and Authorization, we make use of
a common file and a middleware.

The common file , common.login.ts enables us to setup authentication
for different resources that will need login access.

The middleware helps us to authorize a request before moving a
request to the stage in the pipeline.

A request is aborted if it does  not meet with the right pattern