Add Branch Management Feature
We add the branch feature.
This feature maps to the business need of setting up branches.
The value for the maximum number of branches is setup in a config file.

To add a branch, the user must be login an must be an admin.
View right is provided to all users for branch listing and single branch
retrieval.

Operations such as create , update , and deletion of a branch are
exclusive operations to the admin.
Other users cannot perform this operation.

To be able to access a resource, we maintain
a rule that enables or keep our code in check.
This inspired the change of getOne method that
was initially used to get an admin to be changed.