Add Validation for Resources and Reject Duplicate
Each of the endpoints for the resources under control were attached
a validating middleware

The middleware for each of the endpoint checks and ensures that
the needed data to create a resource is provided and in the right
format.

Also, when adding a new record or updating existing one, we want to
ensure it is not a reflection of an already existing item.

Allowing duplicates will lead to less data integrity and data corruption.

To achieve this, we used a library that allows us to be able to check
within any resource for duplicate before adding or updating