Add Integration Test and Make Code Modification
We add an integration test file for each of the request controllers
in the features directory.

Testing the request handles provided hindsight into the part of the
application to modify so as to get everything working well together.

To fix all the issues that were discovered after testing each route ,
modifications were made to code to make them work well

Testing also required that a default img was added in the storage
directory.

This image will be required during testing of customer creation

Also, newer middleware were created such as the one that checks if
a user is a part of the management (isManager)

They (newer middleware) help us to authorize and grant accecss
to some entities at code level when they try to either access a resource
or perform an action on a resource